+++ iot/test/tcp_test.go
@@ -45,7 +45,7 @@ func TestTwoTierTcp(t *testing.T) {
 	sss, _ = iot.GetServerStats(ce.Gurus[0].GetHTTPAddress())
 	fmt.Println("guru stats", sss)
 
-	n := sss.Subscriptions * float32(ce.Gurus[0].Limits.Subscriptions)
+	n := sss.Subscriptions * float64(ce.Gurus[0].Limits.Subscriptions)
 	if n != 1.0 {
 		// t.Errorf("got %v, want %v", n, 1.0)
 	}
@@ -84,7 +84,7 @@ func TestTwoTierTcp(t *testing.T) {
 	fmt.Println("guru stats2", guruStats)
 
 	// the guru gained a subscription because the aide connected to it.
-	n = guruStats.Subscriptions * float32(ce.Gurus[0].Limits.Subscriptions)
+	n = guruStats.Subscriptions * float64(ce.Gurus[0].Limits.Subscriptions)
 	if n != 4.0 {
 		//  t.Errorf("got %v, want %v", n, 4.0)
 	}
